pyproject.toml
[tool.poetry]
name = "dbt-markdoc"
version = "0.1.0"
description = "CLI tool for creating markdown documentation from DBT schema files"
authors = ["Ben Dowling <ben.dowling@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.2"
pydantic = "^1.9.0"
PyYAML = "^6.0"
dbt-core = "^1.0.4"
Jinja2 = "2.11.3"
pylint = "^2.14.5"
mypy = "^0.971"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dbt-markdoc = "dbt-markdoc.cli:run"

[tool.black]
line-length = 120

[tool.pylint.main]
# Include a hint for the correct naming format with invalid-name.
include-naming-hint = true

extension-pkg-whitelist = ["pydantic"]

# Ignore docstrings
disable = [
    "missing-module-docstring",
    "missing-function-docstring",
    "missing-class-docstring"
]

[tool.pylint.format]
# Maximum number of characters on a single line.
max-line-length = 120

[tool.pylint.design]
disable = ["too-few-public-methods"]

[tool.pylint.typecheck]
signature-mutators = [
    "click.decorators.option",
    "click.decorators.argument",
]

[tool.mypy]
ignore_missing_imports = true
strict = true
